requirements.txt
Flask
kubernetes
google-auth
google-api-python-client
pytz
prometheus-flask-exporter